requirements.txt
# Core dependencies for Qualitative Coding Analysis Tool
aiosqlite>=0.17.0
google-generativeai>=0.3.0
fastapi>=0.68.0
uvicorn>=0.24.0
python-docx>=0.8.11
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=5.4.0
anyio>=3.0.0
bleach>=6.0.0
python-json-logger>=2.0.0
tiktoken>=0.5.0
litellm>=1.0.0
python-dotenv>=0.19.0
backoff>=2.2.0

# Graph database
neo4j>=5.0.0

# Export dependencies
pandas>=1.5.0
openpyxl>=3.0.0

# Development dependencies
pytest>=6.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910